move class - domain to application
