+++ src/test/java/co/invest72/investment/application/CalculateInvestmentTest.java
@@ -8,7 +8,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import co.invest72.investment.application.dto.CalculateInvestmentRequest;
-import co.invest72.investment.domain.investment.ExpirationInvestmentFactory;
 import co.invest72.investment.domain.tax.TaxType;
 
 class CalculateInvestmentTest {
@@ -26,7 +25,7 @@ class CalculateInvestmentTest {
 
 	@BeforeEach
 	void setUp() {
-		ExpirationInvestmentFactory investmentFactory = new ExpirationInvestmentFactory();
+		InvestmentFactory investmentFactory = new InvestmentFactory();
 		investmentUseCase = new CalculateExpirationInvestment(investmentFactory);
 		investmentType = FIXED_DEPOSIT.getTypeName();
 		investmentAmount = "1000000";
