+++ src/test/java/co/invest72/investment/domain/InvestmentFactoryTest.java
@@ -7,17 +7,17 @@ import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import co.invest72.investment.application.InvestmentFactory;
 import co.invest72.investment.application.dto.CalculateInvestmentRequest;
 import co.invest72.investment.domain.investment.CompoundFixedDeposit;
 import co.invest72.investment.domain.investment.CompoundFixedInstallmentSaving;
-import co.invest72.investment.domain.investment.ExpirationInvestmentFactory;
 import co.invest72.investment.domain.investment.SimpleFixedDeposit;
 import co.invest72.investment.domain.investment.SimpleFixedInstallmentSaving;
 import co.invest72.investment.domain.tax.TaxType;
 
-class ExpirationInvestmentFactoryTest {
+class InvestmentFactoryTest {
 
-	private ExpirationInvestmentFactory investmentFactory;
+	private InvestmentFactory investmentFactory;
 	private CalculateInvestmentRequest request;
 	private Investment investment;
 
@@ -27,7 +27,7 @@ class ExpirationInvestmentFactoryTest {
 
 	@BeforeEach
 	void setUp() {
-		investmentFactory = new ExpirationInvestmentFactory();
+		investmentFactory = new InvestmentFactory();
 	}
 
 	@Test
